+++ src/LevelManager.cc
@@ -33,12 +33,15 @@
 #include "ignition/gazebo/components/Name.hh"
 #include "ignition/gazebo/components/LevelBuffer.hh"
 #include "ignition/gazebo/components/LevelEntityNames.hh"
+#include "ignition/gazebo/components/LinearVelocity.hh"
+#include "ignition/gazebo/components/LinearVelocitySeed.hh"
 #include "ignition/gazebo/components/MagneticField.hh"
 #include "ignition/gazebo/components/ParentEntity.hh"
 #include "ignition/gazebo/components/Performer.hh"
 #include "ignition/gazebo/components/PerformerLevels.hh"
 #include "ignition/gazebo/components/Pose.hh"
 #include "ignition/gazebo/components/Scene.hh"
+#include "ignition/gazebo/components/Wind.hh"
 #include "ignition/gazebo/components/World.hh"
 
 #include "LevelManager.hh"
@@ -83,14 +86,27 @@ void LevelManager::ReadLevelPerformerInfo()
   this->runner->entityCompMgr.CreateComponent(this->worldEntity,
       components::MagneticField(this->runner->sdfWorld->MagneticField()));
 
+  auto worldElem = this->runner->sdfWorld->Element();
+
+  // Create Wind
+  auto windEntity = this->runner->entityCompMgr.CreateEntity();
+  this->runner->entityCompMgr.CreateComponent(windEntity, components::Wind());
+  this->runner->entityCompMgr.CreateComponent(
+      windEntity, components::WorldLinearVelocity(
+                      this->runner->sdfWorld->WindLinearVelocity()));
+  // Initially the wind linear velocity is used as the seed velocity
+  this->runner->entityCompMgr.CreateComponent(
+      windEntity, components::WorldLinearVelocitySeed(
+                      this->runner->sdfWorld->WindLinearVelocity()));
+
+  this->entityCreator->SetParent(windEntity, this->worldEntity);
+
   // scene
   if (this->runner->sdfWorld->Scene())
   {
     this->runner->entityCompMgr.CreateComponent(this->worldEntity,
         components::Scene(*this->runner->sdfWorld->Scene()));
   }
-
-  auto worldElem = this->runner->sdfWorld->Element();
 
   // TODO(anyone) This should probably go somewhere else as it is a global
   // constant.